chore: added summaries and minor renaming
