+++ Runtime/AvatarCreator/Prefabs/Elements/AvatarPreviewElement.prefab
@@ -12,7 +12,7 @@ GameObject:
   - component: {fileID: 1740648930196985800}
   - component: {fileID: 3027828850030535667}
   m_Layer: 0
-  m_Name: Avatar Preview Element
+  m_Name: AvatarPreviewElement
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
