+++ Runtime/AvatarCreator/Scripts/UI/Elements/AccountCreationElement.cs
@@ -4,13 +4,20 @@ using UnityEngine.UI;
 
 namespace ReadyPlayerMe.AvatarCreator
 {
+    /// <summary>
+    /// This class provides all the functionality required to create a basic Account Creation UI element
+    /// It allows users to input an email address, send an email, and continue without signup.
+    /// </summary>
     public class AccountCreationElement : MonoBehaviour
     {
         [SerializeField] private InputField emailField;
         [SerializeField] private Button sendEmailButton;
         [SerializeField] private Button continueWithoutSignupButton;
 
+        // Event invoked when the "Send Email" button is clicked with the entered email address.
         public UnityEvent<string> onSendEmail;
+
+        // Event invoked when the "Continue Without Signup" button is clicked.
         public UnityEvent onContinueWithoutSignup;
 
         private void Update()
