+++ Runtime/AvatarCreator/Scripts/UI/Elements/GenderSelectElement.cs
@@ -1,4 +1,3 @@
-using System;
 using ReadyPlayerMe.Core;
 using UnityEngine;
 using UnityEngine.Events;
@@ -6,6 +5,10 @@ using UnityEngine.UI;
 
 namespace ReadyPlayerMe.AvatarCreator
 {
+    /// <summary>
+    /// A Unity MonoBehaviour class for gender selection UI.
+    /// Provides buttons to select a gender (male or female) and triggers events upon selection.
+    /// </summary>
     public class GenderSelectElement : MonoBehaviour
     {
         [Header("Buttons")]
@@ -19,23 +22,37 @@ namespace ReadyPlayerMe.AvatarCreator
         [Header("Events")]
         public UnityEvent<OutfitGender> OnGenderSelected;
 
+        /// <summary>
+        /// Adds event listeners when the element is enabled.
+        /// </summary>
         private void OnEnable()
         {
             maleButton.onClick.AddListener(MaleButtonClicked);
             femaleButton.onClick.AddListener(FemaleButtonClicked);
         }
 
+        /// <summary>
+        /// Removes event listeners when the element is disabled.
+        /// </summary>
         private void OnDisable()
         {
             maleButton.onClick.RemoveListener(MaleButtonClicked);
             femaleButton.onClick.RemoveListener(FemaleButtonClicked);
         }
 
+        /// <summary>
+        /// Invoked when the male selection button is clicked.
+        /// Triggers the OnGenderSelected event with the Masculine gender parameter.
+        /// </summary>
         private void MaleButtonClicked()
         {
             OnGenderSelected?.Invoke(OutfitGender.Masculine);
         }
 
+        /// <summary>
+        /// Invoked when the female selection button is clicked.
+        /// Triggers the OnGenderSelected event with the Feminine gender parameter.
+        /// </summary>
         private void FemaleButtonClicked()
         {
             OnGenderSelected?.Invoke(OutfitGender.Feminine);
