+++ Runtime/AvatarCreator/Scripts/UI/Elements/PhotoCaptureElement.cs.meta
