+++ Samples~/AvatarLoadingSamples/AvatarLoading/AvatarLoadingExample.cs
@@ -8,7 +8,7 @@ namespace ReadyPlayerMe.Samples.AvatarLoading
     /// </summary>
     public class AvatarLoadingExample : MonoBehaviour
     {
-        [SerializeField][Tooltip("Set this to the URL or shortcode of the Ready Player Me Avatar you want to load.")]
+        [SerializeField] [Tooltip("Set this to the URL or shortcode of the Ready Player Me Avatar you want to load.")]
         private string avatarUrl = "https://models.readyplayer.me/638df693d72bffc6fa17943c.glb";
 
         private GameObject avatar;
@@ -25,7 +25,7 @@ namespace ReadyPlayerMe.Samples.AvatarLoading
             };
             avatarLoader.LoadAvatar(avatarUrl);
         }
-        
+
         private void OnDestroy()
         {
             if (avatar != null) Destroy(avatar);
